Add react-pdf-node package
