+++ examples/simple/example.pdf
@@ -29,15 +29,12 @@ endobj
 /MediaBox [0 0 595.280029 841.890015]
 >>
 endobj
-
 xref
-0 5
-0000000000 65535 f
+0 50000000000 65535 f
 0000000008 00000 n
 0000000072 00000 n
 0000000128 00000 n
 0000000173 00000 n
-
 trailer
 <<
 /Size 4
