+++ packages/react-pdf-node/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "react-pdf-node",
+  "version": "0.1.0",
+  "main": "src/index.js",
+  "author": "Diego Muracciole <diegomuracciole@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "react": "16.0.0-alpha.4",
+    "react-pdf": "0.1.0"
+  }
+}
